feat(codegen): Support different types of runs in state
