+++ src/seer/automation/autofix/tasks.py
@@ -4,7 +4,6 @@ import random
 from typing import Literal, cast
 
 import sentry_sdk
-import sqlalchemy.sql as sql
 from langfuse import Langfuse
 
 from celery_app.app import celery_app
@@ -40,7 +39,7 @@ from seer.automation.utils import (
     process_repo_provider,
     raise_if_no_genai_consent,
 )
-from seer.db import DbIssueSummary, DbPrIdToAutofixRunIdMapping, DbRunState, Session
+from seer.db import DbPrIdToAutofixRunIdMapping, DbRunState, Session
 
 logger = logging.getLogger(__name__)
 
@@ -94,64 +93,12 @@ def get_autofix_state(
 
 def get_all_autofix_runs_after(after: datetime.datetime):
     with Session() as session:
-        runs = session.query(DbRunState).filter(DbRunState.last_triggered_at > after).all()
+        runs = (
+            session.query(DbRunState)
+            .filter(DbRunState.last_triggered_at > after, DbRunState.type == "autofix")
+            .all()
+        )
         return [ContinuationState.from_id(run.id, AutofixContinuation) for run in runs]
-
-
-def delete_all_runs_before(before: datetime.datetime, batch_size=1000):
-    deleted_count = 0
-    while True:
-        with Session() as session:
-            subquery = (
-                session.query(DbRunState.id)
-                .filter(DbRunState.last_triggered_at < before)
-                .limit(batch_size)
-                .subquery()
-            )
-            count = (
-                session.query(DbRunState)
-                .filter(sql.exists().where(DbRunState.id == subquery.c.id))
-                .delete()
-            )
-            session.commit()
-
-            deleted_count += count
-            if count == 0:
-                break
-            sentry_sdk.metrics.incr(
-                key="autofix_state_TTL_deletion",
-                value=count,
-            )
-
-    return deleted_count
-
-
-def delete_all_summaries_before(before: datetime.datetime, batch_size=1000):
-    deleted_count = 0
-    while True:
-        with Session() as session:
-            subquery = (
-                session.query(DbIssueSummary.group_id)
-                .filter(DbIssueSummary.created_at < before)
-                .limit(batch_size)
-                .subquery()
-            )
-            count = (
-                session.query(DbIssueSummary)
-                .filter(sql.exists().where(DbIssueSummary.group_id == subquery.c.group_id))
-                .delete()
-            )
-            session.commit()
-
-            deleted_count += count
-            if count == 0:
-                break
-            sentry_sdk.metrics.incr(
-                key="issue_summary_TTL_deletion",
-                value=count,
-            )
-
-    return deleted_count
 
 
 @celery_app.task(time_limit=15)
@@ -163,16 +110,6 @@ def check_and_mark_recent_autofix_runs():
     logger.info(f"Got {len(runs)} runs")
     for run in runs:
         check_and_mark_if_timed_out(run)
-
-
-@celery_app.task(time_limit=30)
-def delete_data_for_ttl():
-    logger.info("Deleting old Autofix runs and issue summaries for 90 day time-to-live")
-    before = datetime.datetime.now() - datetime.timedelta(days=90)  # over 90 days old
-    deleted_run_count = delete_all_runs_before(before)
-    deleted_summary_count = delete_all_summaries_before(before)
-    logger.info(f"Deleted {deleted_run_count} runs")
-    logger.info(f"Deleted {deleted_summary_count} summaries")
 
 
 def check_and_mark_if_timed_out(state: ContinuationState):
