+++ tests/automation/autofix/test_autofix_tasks.py
@@ -14,16 +14,13 @@ from seer.automation.autofix.models import (
 )
 from seer.automation.autofix.tasks import (
     check_and_mark_recent_autofix_runs,
-    delete_all_runs_before,
-    delete_all_summaries_before,
-    delete_data_for_ttl,
     get_autofix_state,
     get_autofix_state_from_pr_id,
     run_autofix_create_pr,
     run_autofix_execution,
     run_autofix_root_cause,
 )
-from seer.db import DbIssueSummary, DbPrIdToAutofixRunIdMapping, DbRunState, Session
+from seer.db import DbPrIdToAutofixRunIdMapping, DbRunState, Session
 
 
 class TestGetAutofixState:
@@ -218,166 +215,3 @@ class TestCheckAndMarkRecentAutofixRuns:
         mock_logger.info.assert_any_call("Got 2 runs")
         mock_check_and_mark.assert_has_calls([call(mock_run1), call(mock_run2)])
         assert mock_check_and_mark.call_count == 2
-
-
-class TestDeleteOldAutofixRuns:
-    def test_old_data_is_deleted(self):
-        old_run = DbRunState(
-            last_triggered_at=datetime.datetime.now() - datetime.timedelta(days=100),
-            value="test_value",
-        )
-        old_summary = DbIssueSummary(
-            group_id=1,
-            summary="summary",
-            created_at=datetime.datetime.now() - datetime.timedelta(days=100),
-        )
-        with Session() as session:
-            session.add(old_run)
-            session.add(old_summary)
-            session.commit()
-
-        before_date = datetime.datetime.now() - datetime.timedelta(days=90)
-        deleted_run_count = delete_all_runs_before(before_date)
-        deleted_summary_count = delete_all_summaries_before(before_date)
-
-        with Session() as session:
-            remaining_run = session.query(DbRunState).filter(DbRunState.id == old_run.id).first()
-            remaining_summary = (
-                session.query(DbIssueSummary)
-                .filter(DbIssueSummary.group_id == old_summary.group_id)
-                .first()
-            )
-        assert remaining_run is None
-        assert remaining_summary is None
-        assert deleted_run_count == 1
-        assert deleted_summary_count == 1
-
-    def test_data_not_deleted_within_ttl(self):
-        recent_run = DbRunState(
-            last_triggered_at=datetime.datetime.now() - datetime.timedelta(days=80),
-            value="test_value",
-        )
-        recent_summary = DbIssueSummary(
-            group_id=1,
-            summary="summary",
-            created_at=datetime.datetime.now() - datetime.timedelta(days=80),
-        )
-        with Session() as session:
-            session.add(recent_run)
-            session.add(recent_summary)
-            session.commit()
-
-        before_date = datetime.datetime.now() - datetime.timedelta(days=90)
-        deleted_run_count = delete_all_runs_before(before_date)
-        deleted_summary_count = delete_all_summaries_before(before_date)
-
-        with Session() as session:
-            remaining_run = session.query(DbRunState).filter(DbRunState.id == recent_run.id).first()
-            remaining_summary = (
-                session.query(DbIssueSummary)
-                .filter(DbIssueSummary.group_id == recent_summary.group_id)
-                .first()
-            )
-        assert remaining_run is not None
-        assert remaining_summary is not None
-        assert remaining_run.last_triggered_at > before_date
-        assert remaining_summary.created_at > before_date
-        assert deleted_run_count == 0
-        assert deleted_summary_count == 0
-
-    def test_batch_delete(self):
-        old_runs = [
-            DbRunState(
-                last_triggered_at=datetime.datetime.now() - datetime.timedelta(days=100),
-                value="test_value",
-            )
-            for _ in range(25)
-        ]
-        old_summaries = [
-            DbIssueSummary(
-                group_id=i,
-                summary="summary",
-                created_at=datetime.datetime.now() - datetime.timedelta(days=100),
-            )
-            for i in range(35)
-        ]
-        with Session() as session:
-            session.bulk_save_objects(old_runs)
-            session.bulk_save_objects(old_summaries)
-            session.commit()
-
-        before_date = datetime.datetime.now() - datetime.timedelta(days=90)
-        deleted_run_count = delete_all_runs_before(before_date, batch_size=10)
-        deleted_summary_count = delete_all_summaries_before(before_date, batch_size=10)
-
-        with Session() as session:
-            remaining_runs = (
-                session.query(DbRunState).filter(DbRunState.last_triggered_at < before_date).count()
-            )
-            remaining_summaries = (
-                session.query(DbIssueSummary)
-                .filter(DbIssueSummary.created_at < before_date)
-                .count()
-            )
-        assert remaining_runs == 0
-        assert remaining_summaries == 0
-        assert deleted_run_count == 25
-        assert deleted_summary_count == 35
-
-    def test_delete_old_autofix_runs_task(self):
-        old_run = DbRunState(
-            last_triggered_at=datetime.datetime.now() - datetime.timedelta(days=100),
-            value="test_value",
-        )
-        old_summary = DbIssueSummary(
-            group_id=1,
-            summary="summary",
-            created_at=datetime.datetime.now() - datetime.timedelta(days=100),
-        )
-        with Session() as session:
-            session.add(old_run)
-            session.add(old_summary)
-            session.commit()
-
-        # Celery task
-        delete_data_for_ttl()
-
-        with Session() as session:
-            remaining_run = session.query(DbRunState).filter(DbRunState.id == old_run.id).first()
-            remaining_summary = (
-                session.query(DbIssueSummary)
-                .filter(DbIssueSummary.group_id == old_summary.group_id)
-                .first()
-            )
-        assert remaining_run is None
-        assert remaining_summary is None
-
-    def test_delete_with_mapped_pr(self):
-        old_run = DbRunState(
-            id=789,
-            last_triggered_at=datetime.datetime.now() - datetime.timedelta(days=100),
-            value="test_value",
-        )
-        mapped_pr = DbPrIdToAutofixRunIdMapping(
-            provider="github",
-            pr_id=123,
-            run_id=789,
-        )
-        with Session() as session:
-            session.add(old_run)
-            session.add(mapped_pr)
-            session.commit()
-
-        before_date = datetime.datetime.now() - datetime.timedelta(days=90)
-        deleted_count = delete_all_runs_before(before_date)
-
-        with Session() as session:
-            remaining_run = session.query(DbRunState).filter(DbRunState.id == old_run.id).first()
-            remaining_pr = (
-                session.query(DbPrIdToAutofixRunIdMapping)
-                .filter(DbPrIdToAutofixRunIdMapping.run_id == old_run.id)
-                .first()
-            )
-        assert remaining_run is None
-        assert remaining_pr is None
-        assert deleted_count == 1
